Took the code from sneek rev. 172 and updated ours to stay in sync.
A bugfix that caused es not to wait for di in sneek.
Lowering the FS FS-USB and ES priority levels
When a file open fails in write mode, retry the action with the create_always flag set.

The code is changed but not tested yet. All I can tell is that it compiles.